Fix bad conversions when converting B to bin bytes in AS mode
When converting from bytes to binary bytes such as B-KiB, B-MiB, B-GiB,
the calculation ended up being wrong and was missing the rest of the
value. This was because the amount held by bytes was an integer and not
a floating point/double data type.

+++ include/bytesize.h
@@ -40,7 +40,7 @@ static const char *BYTE[SIZE] = {
 };
 
 typedef struct Byte {
-  int amt;
+  double amt;
   char* unit;
   long int scaling;
 } Byte;
